+++ prqlc/prql-compiler/tests/integration/sql.rs
@@ -2004,6 +2004,37 @@ fn test_distinct_on_01() {
 }
 
 #[test]
+fn test_group_take_n() {
+    assert_display_snapshot!((compile(r###"
+    prql target:sql.postgres
+
+    from employees
+    group department (
+      sort age
+      take 2
+    )
+    "###).unwrap()), @r###"
+    WITH table_0 AS (
+      SELECT
+        *,
+        ROW_NUMBER() OVER (
+          PARTITION BY department
+          ORDER BY
+            age
+        ) AS _expr_0
+      FROM
+        employees
+    )
+    SELECT
+      *
+    FROM
+      table_0
+    WHERE
+      _expr_0 <= 2
+    "###);
+}
+
+#[test]
 fn test_distinct_on_02() {
     assert_display_snapshot!((compile(r###"
     prql target:sql.duckdb